review: use uint64 in stats expectations.

+++ mixer/test/client/check_report/check_report_test.go
@@ -199,7 +199,7 @@ const reportAttributesOkPost = `
 `
 
 // Stats in Envoy proxy.
-var expectedStats = map[string]int{
+var expectedStats = map[string]uint64{
 	// Policy check stats
 	"http_mixer_filter.total_check_calls":             2,
 	"http_mixer_filter.total_check_cache_hits":        0,
